refactor(IN-311): coderabbitai 수정 반영
