+++ src/pages/me/alarm/ui/MyAlarmPage.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { Switch } from '@/shared/ui/shadcn/switch'
+import { Switch } from '@/shared/ui/shadcn/Switch'
 
 export function MyAlarmPage() {
   return (
@@ -23,7 +23,7 @@ export function MyAlarmPage() {
                 결제 관련 중요한 알림
               </span>
             </div>
-            <Switch />
+            <Switch aria-label='결제 알림 수신' />
           </div>
 
           <div className='flex items-end justify-between gap-8'>
@@ -35,7 +35,7 @@ export function MyAlarmPage() {
                 결제 완료, 실패, 다음 결제 예정일을 알려드려요
               </span>
             </div>
-            <Switch />
+            <Switch aria-label='결제 관련 알림 수신' />
           </div>
         </div>
         {/* 마케팅 알림 */}
@@ -49,7 +49,7 @@ export function MyAlarmPage() {
                 새 기능 및 이벤트 안내 (선택)
               </span>
             </div>
-            <Switch />
+            <Switch aria-label='마케팅 알림 수신' />
           </div>
 
           <div className='flex items-end justify-between gap-8'>
@@ -61,7 +61,7 @@ export function MyAlarmPage() {
                 플랫폼의 새로운 기능을 가장 먼저 알려드려요
               </span>
             </div>
-            <Switch />
+            <Switch aria-label='새 기능 안내 수신' />
           </div>
 
           <div className='flex items-end justify-between gap-8'>
@@ -73,7 +73,7 @@ export function MyAlarmPage() {
                 특별 할인이나 이벤트 정보를 보내드려요
               </span>
             </div>
-            <Switch />
+            <Switch aria-label='이벤트 및 혜택 수신' />
           </div>
         </div>
       </div>
